tp2/app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
neo4j==5.15.0
pandas==2.1.3
python-dotenv==1.0.0
requests==2.31.0

